Introduce network protocol version 9.
This adds a 'game ID' to the end of the message.

The ID is a randomly chosen 32-bit value, which clients can use to
prevent against problems seen when secondary game controllers are
started on a network during a game.

Clients should cache the game ID they first observe and ignore messages
for game IDs other than this.

If they don't see any game controller message for some period of time
(say 10 seconds) they may forget this ID and take the next observed one.
This will allow the game controller to be restarted safely.

Using a game ID in this way is much simpler than using MAC address
filtering on the network, as was done during the Iran Open between
FUmanoids and MRL (kid size).

Bold Hearts also played a game in Brazil during which another team
accidentally started a game controller on our field's wifi. This caused
our robots to flip between playing and ready states.

With this patch, message of versions 7, 8 and 9 are all broadcast
simultaneously.