Remove static application state.
ActionHandler (was EventHandler) and timeline portion of Log moved into
new Game class, which has instance lifetime.

Action objects are passed to Game.apply which mutates its internal state.

Game class is the single source for up-to-date game state.

GameState.message field removed. New TimelineEntry class couples a string
description of a stored state along with the clone of that state.

Renamed GCActin to Action.

Actions no longer have a type. The type (which is really the trigger,
such as User/Clock/Network) is provided at the time the Action is applied
to the Game.