Make more widespread use of Pair<T> and UIOrientation.
The use of arrays to hold per-team objects can be confusing. What is the
index (side/color/something else)? Also, it's prone to bounds errors.
Using the Pair<T> class means that a dereference is explicit (via UISide
and TeamColor enums), and that the mapping between colour and UI side may
be modified at runtime without having to update each and every array
throughout the application.